Rename hooks to get FullStructure
