--- src/utils/hooks/useGetFullReviewStructure/index.tsx
@@ -1,5 +1,0 @@
-import useGetFullReviewStructure from './useGetFullReviewStructure'
-import useGetFullReviewStructureAsync from './useGetFullReviewStructureAsync'
-
-export default useGetFullReviewStructure
-export { useGetFullReviewStructureAsync }
