+++ src/utils/hooks/useGetReviewStructureForSection/index.tsx
@@ -1,0 +1,5 @@
+import useGetReviewStructureForSections from './useGetReviewStructureForSection'
+import useGetReviewStructureForSectionAsync from './useGetReviewStructureForSectionAsync'
+
+export default useGetReviewStructureForSections
+export { useGetReviewStructureForSectionAsync as useGetFullReviewStructureAsync }
